BE-53-Feat-STOMP-API

+++ src/main/java/org/tenten/tentenstomp/domain/trip/dto/response/TripPathInfoMsg.java
@@ -14,7 +14,7 @@ public record TripPathInfoMsg(
 
 ) {
     public record PathInfo(
-        Long price,
+        Integer price,
         Double totalDistance,
         Long totalTime
     ){}
